+++ editButton.js
@@ -1,0 +1,13 @@
+import { comments } from "./api.js";
+import { renderComments } from "./render.js";
+
+export const editButtonListners = () => {
+    const editButtonElements = document.querySelectorAll(".edit");
+    for (const editButtonElement of editButtonElements) {
+      editButtonElement.addEventListener("click", (event) => {
+        event.stopPropagation();
+        comments[editButtonElement.dataset.index].isEdit = false;
+        renderComments({comments});
+      })
+    }
+  }
