+++ enableButton.js
@@ -1,0 +1,5 @@
+export const enableButtonOnInput = (InputElement, buttonElement) => {
+    InputElement.addEventListener("input", () => {
+       buttonElement.disabled = false;
+    });
+};
