+++ massangeS.js
@@ -1,0 +1,6 @@
+import { massageSendButton } from "./api.js";
+
+export const massageSendListners = (massageSendButton, {addComment}) => {
+  return massageSendButton.addEventListener("click", addComment);
+  
+}
