+++ setInput.js
@@ -1,0 +1,7 @@
+export const setInputValidation = (inputElement) => {
+    inputElement.oninput = () => {
+        if (inputElement.value.charAt(0) === ' ') {
+            inputElement.value = '';
+        }
+    };
+};
